+++ swing/framework-ui/src/main/java/is/codion/swing/framework/ui/EntityEditorPanel.java
@@ -181,6 +181,13 @@ public class EntityEditorPanel extends JPanel {
 	}
 
 	/**
+	 * @return the underlying editor
+	 */
+	public final SwingEntityEditor editor() {
+		return editor;
+	}
+
+	/**
 	 * @return the {@link InputFocus} instance, managing the input focus
 	 */
 	public final InputFocus focus() {